Updated SellerSimpleDto to use real values

+++ ProductService/src/main/java/com/product/service/dto/seller/SellerSimpleDto.java
@@ -3,7 +3,7 @@ package com.product.service.dto.seller;
 public record SellerSimpleDto(
         Long id,
         String name,
-        Integer newProductsLastMonth,
-        Integer newSalesLastMonth
+        long newProductsLastMonth,
+        long newSalesLastMonth
 ) {
 }
